requirements.txt
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
Django==4.0.5
django-browser-reload==1.6.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-filter==21.1
django-modelcluster==5.3
django-redis==5.2.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
mypy-extensions==0.4.3
openpyxl==3.0.10
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.9
wagtail==2.16
webencodings==0.5.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
